complete design and workaround args by HeuteFinden

+++ code/src/main/java/bob/belu/HeuteFinden.java
@@ -7,37 +7,58 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.LinkedHashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class HeuteFinden {
 
+	private static boolean verbose = false;
+
 	public static void main(String[] args) throws IOException {
 		String path = null;
+		String use = null;
+		List<String> vars = new LinkedList<>();
 		String out = null;
 		for (String arg : args) {
 			if (arg.matches("-[fF][iI][lL][eE]=.*")) {
 				path = arg.substring(6);
+			} else if (arg.matches("-[uU][sS][eE]=.*")) {
+				use = arg.substring(5);
+			} else if (arg.toLowerCase().startsWith("-var=") && -1 < arg.indexOf(':')) {
+				// arg.matches("-[vV][aA][rR]=.*")
+				vars.add(arg.substring(5).replaceAll("`", "\""));
 			} else if (arg.matches("-[oO][uU][tT]=.*")) {
 				out = arg.substring(5);
+			} else if (arg.matches("-[vV][eE][rR][bB][oO][sS][eE]=.*")) {
+				verbose = Boolean.parseBoolean(arg.substring(9));
+			} else {
+				System.out.println("argument unknown: " + mask(arg));
 			}
 		}
 		if (null == path) {
 			System.err.println("Datei fehlt :(");
+		} else if (null == use) {
+			System.err.println("Parameter \"use\" fehlt :(");
+		} else if (0 == vars.size()) {
+			System.err.println("Parameter \"var\" fehlt :(");
 		} else {
-			new HeuteFinden(path, out);
+			new HeuteFinden(path, use, vars, out);
 		}
 	}
 
-	public String temp = null;
+	private Map<String,String> values = new LinkedHashMap<>();
 
-	public String icon = null;
+	public HeuteFinden(final String path, String use, List<String> vars, String out) throws IOException {
 
-	public String cond = null;
+		boolean overwrite = true;
 
-	public HeuteFinden(final String path, String out) throws IOException {
-		System.out.println("lese " + path);
 		// Laden
+		System.out.println("lese " + path);
 		StringBuffer buffer = new StringBuffer();
 		BufferedReader file = new BufferedReader(new FileReader(path));
 		String line = null;
@@ -46,26 +67,26 @@ public class HeuteFinden {
 		}
 		file.close();
 
-		String value = search("Infoblock", "(<li class=\"day current.+?</li>)", buffer.toString());
+		// "(<li class=\"day current.+?</li>)"
+		String value = search("use", use, buffer.toString());
 
-		// Icon
-		// http://vortex.accuweather.com/adc2010/images/slate/icons/16.svg
-		icon = search("Symbol", "icon i-(\\d+)-xl", value);
+		for (String v: vars) {
+			String varName = v.substring(0, v.indexOf(':'));
+			String varRegex = v.substring(v.indexOf(':') + 1);
+			System.out.println("suche " + varName + " mit \"" + mask(varRegex) + "\"");
+		
+			String varValue = search(varName, varRegex, value);			
+			System.out.println(varName + " ist " + varValue);
 
-		// Temperatur
-		// 9°
-		temp = search("Temperatur", "<span class=\"large-temp\">(.*)&deg;\n</span>", value);
-
-		// Bedingungen
-		// Windig, nachts einige Schauer
-		cond = search("Bedingungen", "<span class=\"cond\">(.*)\n</span>", value);
+			values.put(varName, varValue);
+		}
 
 		// Ausgabe
 		if (null == out) {
 			System.out.println("Parameter \"out\" fehlt. Keine Ausgabe!");
 		} else {
-			File f = new File(out, "beluData.js");
-			if (f.exists()) {
+			File f = new File(out, "beluWeather.js");
+			if (f.exists() && !overwrite) {
 				String ts = new SimpleDateFormat("yyyyMMdd_HHmmss").format(new Date());
 				File backup = new File(f.getAbsolutePath().replaceAll("\\.js", "_" + ts + "\\.js"));
 				f.renameTo(backup);
@@ -74,34 +95,47 @@ public class HeuteFinden {
 			System.out.println("schreibe \"" + f.getAbsolutePath() + "\"");
 			FileWriter writer = new FileWriter(f);
 			writer.write("// Daten vom Wetterdienst\n");
-			writer.write("var weatherTemp = " + temp + ";\n");
-			writer.write("var weatherIcon = " + icon + ";\n");
-			writer.write("var weatherCond = \"" + cond + "\";\n");
+			for (Entry<String,String> e: values.entrySet()) {
+				writer.write("var " + e.getKey() + " = \"" + e.getValue() + "\";\n");
+			}
 			writer.close();
 		}
 
 	}
 
 	private String search(String label, String pattern, String text) {
-		String t = "----- " + label + ": " + pattern.replaceAll("\\r", "\\\\r").replaceAll("\\n", "\\\\n") + " -----";
-		System.out.println(t);
+		String t = "----- " + label + ": " + mask(pattern) + " -----";
+		println(t);
 
 		String x = null;
 		if (null != text) {
-			Pattern p = Pattern.compile(pattern);
+			String p1 = pattern.replaceAll("\\\\\\\\d", "\\\\d");
+			Pattern p = Pattern.compile(p1);
 			Matcher m = p.matcher(text);
 			if (m.find()) {
 				x = m.group(1)
 					.replaceAll("[\\s][\\s]+", " ")
 					.replaceAll("<([^<>]+)>", "\n<$1>")
 					.trim();
-				System.out.println(x);
+				println(x);
 			} else {
-				System.out.println("Nicht gefunden :(");
+				println("Nicht gefunden :(");
 			}
 		}
-		System.out.println(new String(new char[t.length()]).replace("\0", "-"));
+		
+		println(new String(new char[t.length()]).replace("\0", "-"));
+
 		return x;
+	}
+
+	private static void println(String value) {
+		if (verbose) {
+			System.out.println(value);
+		}
+	}
+
+	private static String mask(String value) {
+		return (null == value ? null : value.replaceAll("\\r", "\\\\r").replaceAll("\\n", "\\\\n"));
 	}
 
 }
